Move read data to
