+++ src/dave_data/__init__.py
@@ -3,12 +3,13 @@ __version__ = "v0.0.1"
 from dave_data.geometry.area_to_polygon import federal_state_to_polygon
 from dave_data.geometry.area_to_polygon import postalcode_to_polygon
 from dave_data.geometry.area_to_polygon import town_to_polygon
+
 from .core import compute
 from .core import get_data
 from .datapool.osm_request import osm_request
-from .datapool.read_data import read_federal_states
-from .datapool.read_data import read_nuts_regions
-from .datapool.read_data import read_postal
+from dave_data.io.read_data import read_federal_states
+from dave_data.io.read_data import read_nuts_regions
+from dave_data.io.read_data import read_postal
 
 __all__ = [
     "postalcode_to_polygon",
